ajustes

+++ src/business-rules/customers/handlers/delete-customer.handler.ts
@@ -1,12 +1,23 @@
-import { Injectable } from '@nestjs/common';
-import { CustomerRepository } from '@domain/repositories';
+import { IDeleteCustomerHandler } from './idelete-customer.handler';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ICustomerRepository,
+  ICustomerRepositoryName,
+} from '@domain/repositories';
 import { IDeleteCustomerRequest } from '../requests';
 
 @Injectable()
-export class DeleteCustomerHandler {
-  constructor(private repository: CustomerRepository) {}
+export class DeleteCustomerHandler implements IDeleteCustomerHandler {
+  constructor(
+    @Inject(ICustomerRepositoryName) private repository: ICustomerRepository,
+  ) {}
 
   async execute(request: IDeleteCustomerRequest): Promise<void> {
+    const customer = await this.repository.findById(request.id);
+    if (!customer) {
+      throw new NotFoundException('Customer not found');
+    }
+
     await this.repository.delete(request.id);
   }
 }
